+++ src/Command/RegisterCommand.php
@@ -45,7 +45,8 @@ class RegisterCommand extends Command
 			->setName('dbchanger:register')
 			->setDescription('Register DbChange and prepare it for further installation')
 			->addArgument('code', InputArgument::REQUIRED, 'DbChange code to be registered.')
-			->addOption('debug', 'd', InputOption::VALUE_NONE, 'y');
+			->addOption('debug', 'd', InputOption::VALUE_NONE, 'y')
+			->addOption('overwrite', 'o', InputOption::VALUE_NONE, 'y');
 	}
 
 	protected function execute(InputInterface $input, OutputInterface $output)
@@ -54,7 +55,7 @@ class RegisterCommand extends Command
 
 		try {
 			$dbChange = $this->loader->loadDbChangeFromInputDirectory($this->manager->getGroups(), $dbChangeCode);
-			$this->manager->registerDbChange($dbChange, $input->getOption('debug'));
+			$this->manager->registerDbChange($dbChange, $input->getOption('debug'), $input->getOption('overwrite'));
 			$output->writeln(sprintf('DbChange %s successfully registered.', $dbChangeCode));
 		} catch (DbChangeException $e) {
 			$output->writeln($e->getMessage());