+++ RELEASE_NOTES.md
@@ -1,3 +1,12 @@
+0.5.3 - 2018/07/03
+=================
+* Added overwrite mode for command `register` which will overwrite existing DbChanges with the same code (providing no pending installation is detected)
+```
+php bin/console.php dbchanger:register 12345
+php bin/console.php dbchanger:register 12345 -o
+```
+
+
 0.5.2 - 2018/05/29
 =================